rename News2.vue component to NewsFromAPI.vue
